quiet ping. portHtmlParser refactorized
