+++ portScanner.py
@@ -17,83 +17,9 @@ import os
 import re
 import sys
 import time
+import subprocess
 from threading import Thread
-from bs4 import BeautifulSoup
-
-class Port:
-	"""
-	This class defines a port info.
-	"""
-	def __init__(self, portNum, service, details):
-		"""
-		:param portNum: port number
-		:param service: port service
-		:param details: port service details
-		"""
-		self.portNum = portNum
-		self.service = service
-		self.details = details
-
-
-def trim(s, finalDelimiter):
-	"""
-	This function trims a string from position 0 to the delimiter passed as a parameter
-
-	:param s: string with the HTML code
-	:param finalDelimiter: HTML tag or tags indicating how far to trim
-
-	:return obj 'str': returns a trimmed version of param "s"
-	"""
-
-	# End position
-	delimiter = s.index(finalDelimiter) + len(finalDelimiter)
-	# Return trimmed HTML code
-	return s[delimiter:]
-
-
-def extract(s, startDelimiter, finalDelimiter, ommitSpaces=True):
-	"""
-	This function extracts info between HTML tags.
-
-	:param s: string with the HTML code
-	:param startDelimiter: HTML tag or tags indicating where the valid info starts
-	:param finalDelimiter: HTML tag or tags indicating where the valid info ends
-	:param ommitSpaces (optional, default=True): indicates if spaces should be removed or not
-
-	:return obj 'str': returns a trimmed version of param "s"
-	"""
-
-	# Start and end positions
-	start = s.index(startDelimiter) + len(startDelimiter)
-	end = s.index(finalDelimiter)
-	# Return valid info
-	if ommitSpaces:
-		return s[start:end].strip(' \n\t')
-	else:
-		return s[start:end].strip('\n\t')
-
-
-def getPortFromHTML(html):
-	"""
-	This function extracts port info and trims HTML to avoid HTML tags colitions.
-
-	:param html: string with the HTML code of the row containing port info
-
-	:return obj 'Port': returns a new Port object with the info extracted from the HTML
-	"""
-
-	# Extract port number and trim HTML
-	portNum = extract(html, '<tr><td>', '</td>')
-	html = trim(html, '</td>')
-	# Extract port service and trim HTML
-	service = extract(html, '<td>', '</td>')
-	html = trim(html, '</td>')
-	# Extract port service details
-	details = extract(html, '<td>', '</td></tr>', ommitSpaces=False)
-
-	# Return a new port with the obtained info
-	return Port(portNum, service, details)
-
+import portHtmlParser
 
 def binarySearch(targetPortNum, start, end):
 	"""
@@ -126,7 +52,6 @@ def binarySearch(targetPortNum, start, end):
 			# Look for the port number in the left.
 			return binarySearch(targetPortNum, start, mid - 1)
 
-
 def scanPort(hostIP, portNum):
 	"""
 	This function checks if a specified port is listening.
@@ -139,6 +64,7 @@ def scanPort(hostIP, portNum):
 
 	# Create socket
 	s = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
+	#TODO: HANDLE WITH "[Errno 24] Too many open files"
 
 	# Check if port is listening
 	if s.connect_ex((hostIP, portNum)) == 0:
@@ -148,53 +74,47 @@ def scanPort(hostIP, portNum):
 			print("  - %s (%s) --> %s\n" % (port.service, port.portNum, port.details))
 		s.close()
 
+if __name__ == "__main__":
 
-# All well-known ports (1-1024)
-allPorts = []
-# IP address pattern
-pattern = re.compile("^\d{1,3}\.\d{1,3}\.\d{1,3}\.\d{1,3}$")
+	# IP address pattern
+	pattern = re.compile("^\d{1,3}\.\d{1,3}\.\d{1,3}\.\d{1,3}$")
 
-while(True):
-	# Ask the user for an IP address
-	hostIP = str(raw_input("\n Introduce the host's IP address: "))
+	while(True):
+		# Ask the user for an IP address
+		hostIP = str(raw_input("\n Introduce the host's IP address: "))
 
-	# Check if the user's input is a valid IP address
-	if pattern.match(hostIP):
-		time.sleep(1.2)
-		print("\n Checking host connection...\n")
-		time.sleep(2)
-		# Check if host is alive (ping)
-		pingCheck = os.system('ping -c 3 ' + hostIP)
-		if pingCheck == 0:
+		# Check if the user's input is a valid IP address
+		if pattern.match(hostIP):
 			time.sleep(1.2)
-			print("\n\n SUCCESS: Host is up")
-			break
+			print("\n Checking host connection...\n")
+			time.sleep(2)
+			# Check if host is alive (ping)
+			p = subprocess.Popen(["ping", "-q", "-c", "3", hostIP], stdout = subprocess.PIPE, stderr=subprocess.PIPE) # discard stdout, stderr
+			lostPackets = p.wait()
+			if not lostPackets:
+				time.sleep(1.2)
+				print("\n\n SUCCESS: Host is up")
+				break
+			else:
+				print("\n ** Host is not alive")
+				sys.exit("\n Execution stopped\n")
 		else:
-			print("\n ** Host is not alive")
-			sys.exit("\n Execution stopped\n")
-	else:
-		print("\n\t** ERROR: not valid IP address")
-		again = str(raw_input("\n Try again? (y/n): "))
-		if again == "y" or again == "Y":
-			continue
-		elif again == "n" or again == "N":
-			sys.exit("\n Execution stopped\n")
-		else:
-			sys.exit("\n\t** ERROR: not valid option %s\n" % again)
+			print("\n\t** ERROR: not valid IP address")
+			again = str(raw_input("\n Try again? (y/n): "))
+			if again == "y" or again == "Y":
+				continue
+			elif again == "n" or again == "N":
+				sys.exit("\n Execution stopped\n")
+			else:
+				sys.exit("\n\t** ERROR: not valid option %s\n" % again)
 
 
-print("\n\n *************** LISTENING PORTS *************** \n")
+	print("\n\n *************** LISTENING PORTS *************** \n")
 
-# Get HTML
-html = urllib2.urlopen('http://web.mit.edu/rhel-doc/4/RH-DOCS/rhel-sg-en-4/ch-ports.html').read()
-soup = BeautifulSoup(html, 'html.parser')
+	# All well-known ports (1-1024)
+	allPorts = portHtmlParser.parsePorts()
 
-# Parse HTML to find ports info
-tableRows = soup.find('table', class_='CALSTABLE').find('tbody').findChildren('tr')
-for row in tableRows:
-	allPorts.append(getPortFromHTML(str(row)))
-
-# Check which well-known ports are listening
-for portNum in range(1,1025):
-	thread = Thread(target = scanPort, args = (hostIP, portNum,))
-	thread.start()
+	# Check which well-known ports are listening
+	for portNum in range(1,1025):
+		thread = Thread(target = scanPort, args = (hostIP, portNum))
+		thread.start()
